requirements.txt
Django==3.2.4
Pillow==7.2.0
gunicorn==20.1.0
whitenoise==5.2.0
sentry-sdk==1.1.0
requests==2.25.1
django-redis==5.0.0
pytz==2021.1
celery==4.4.7
celery-redbeat==2.0.0
psycopg2-binary==2.8.6
arrow==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
django-environ==0.4.5
django-storages==1.11.1
django-cors-headers==3.7.0
django-braces==1.14.0
django-extensions==3.1.3
django-post-office==3.5.3
django-crispy-forms==1.11.2
django-allauth==0.44.0
boto3==1.17.86
boto3-stubs==1.17.86
django-import-export==2.5.0
honeybadger==0.6.0
django-ses==2.0.0
djangoql==0.14.5
flake8==3.9.2
